refactor: bug fix && bit update on uiux
